pdb.rs: fix `DeviceSQLString::parse_long_utf16le`
previous code relied on kaitai docs that theorized the presence of
two nulls at the end of these strings. The kaitai parser definition
does not check this so the theory has never been proven in practice.
Its more likely that the string just ends without any trailing null
bytes and there is likely garbage which must not be interpreted.