build(cargo): Bump version to 0.2.0
Since v0.1.0 we switched from `nom` to `binrw` and added support for
parsing `*SETTING.DAT` files, so it's really more than just a bug fix
release. Although the switch to `binrw` is an API change, the major
we don't want to bump the version `1` in order to emphasize that the
API is not finalized and subject to changes.