Correct row count fields
The field num_rows_small was incorrectly interpreted, and is actually a
field that spans up to 13 bits. This is followed by 11 bits that count
the number of valid rows, previously known as unknown3/unknown4.

Fixing this, it becomes apparent that the 13 bit row count is the full
row count, and num_rows_large is not actually a row count at all. I
renamed that field to clarify.